Backend/requirements.txt
asgiref==3.7.2
click==8.1.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-choices-field==2.2.2
django-environ==0.10.0
django-import-export==3.2.0
et-xmlfile==1.1.0
graphql-core==3.2.3
h11==0.14.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.199.3
strawberry-graphql-django==0.16.0
tablib==3.5.0
typing_extensions==4.7.1
uvicorn==0.23.2
xlrd==2.0.1
xlwt==1.3.0
